% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_discrete_Y}
\alias{check_discrete_Y}
\title{check if Y is discrete}
\usage{
check_discrete_Y(Y)
}
\arguments{
\item{Y}{matrix of input outcomes Y}
}
\value{
Y
}
\description{
Checks the input data as discrete outcome
}
\keyword{internal}
