% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_eval}
\alias{check_eval}
\title{check evaluation for margins}
\usage{
check_eval(eval)
}
\arguments{
\item{eval}{string, evaluation points for margins}
}
\value{
eval
}
\description{
Checks the input data of eval in margins
}
\keyword{internal}
