% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_honesty}
\alias{check_honesty}
\title{check honesty}
\usage{
check_honesty(honesty)
}
\arguments{
\item{honesty}{logical, if TRUE honest forest is built using 50:50 data split}
}
\value{
honesty
}
\description{
Checks the input for honesty
}
\keyword{internal}
