% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_honesty_fraction}
\alias{check_honesty_fraction}
\title{check honesty fraction}
\usage{
check_honesty_fraction(honesty.fraction, honesty)
}
\arguments{
\item{honesty.fraction}{scalar, share of the data set aside to estimate the effects (default is 0.5)}

\item{honesty}{logical, if data should be split into train and honest sample}
}
\value{
honesty.fraction
}
\description{
Checks the input data of honesty.fraction in orf
}
\keyword{internal}
