% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_importance}
\alias{check_importance}
\title{check importance}
\usage{
check_importance(importance)
}
\arguments{
\item{importance}{logical, if TRUE variable importance is conducted}
}
\value{
importance
}
\description{
Checks the input for importance
}
\keyword{internal}
