% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_inference}
\alias{check_inference}
\title{check inference}
\usage{
check_inference(inference)
}
\arguments{
\item{inference}{logical, if TRUE the weight based inference is conducted}
}
\value{
inference
}
\description{
Checks the input for inference
}
\keyword{internal}
