% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_latex}
\alias{check_latex}
\title{check latex}
\usage{
check_latex(latex)
}
\arguments{
\item{latex}{logical, TRUE if latex summary should be generated}
}
\value{
latex
}
\description{
Checks the input for latex
}
\keyword{internal}
