% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_min_node_size}
\alias{check_min_node_size}
\title{check min.node.size}
\usage{
check_min_node_size(min.node.size, X)
}
\arguments{
\item{min.node.size}{scalar, minimum node size}

\item{X}{matrix of input features X}
}
\value{
min.node.size
}
\description{
Checks the input data of min.node.size
}
\keyword{internal}
