% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_mtry}
\alias{check_mtry}
\title{check mtry}
\usage{
check_mtry(mtry, X)
}
\arguments{
\item{mtry}{scalar, number of randomly selected features}

\item{X}{matrix of input features X}
}
\value{
mtry
}
\description{
Checks the input data of mtry
}
\keyword{internal}
