% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_newdata}
\alias{check_newdata}
\title{check newdata}
\usage{
check_newdata(new_data, X)
}
\arguments{
\item{new_data}{matrix X containing the observations to predict}

\item{X}{matrix of input features X}
}
\description{
Checks the input for newdata for predict.orf
}
\keyword{internal}
