% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_nmin}
\alias{check_nmin}
\title{check nmin}
\usage{
check_nmin(nmin, X)
}
\arguments{
\item{nmin}{scalar, minimum node size}

\item{X}{matrix of input features X}
}
\value{
nmin
}
\description{
Checks the input data of mtry
}
