% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_num_trees}
\alias{check_num_trees}
\title{check num.trees}
\usage{
check_num_trees(num.trees)
}
\arguments{
\item{num.trees}{scalar, number of trees to be estimated}
}
\value{
num.trees
}
\description{
Checks the input data of num.trees
}
\keyword{internal}
