% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_replace}
\alias{check_replace}
\title{check replace}
\usage{
check_replace(replace)
}
\arguments{
\item{replace}{logical, if TRUE bootstrapping, if FALSE subsampling}
}
\value{
replace
}
\description{
Checks the input for replace
}
\keyword{internal}
