% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_sample_fraction}
\alias{check_sample_fraction}
\title{check sample.fraction}
\usage{
check_sample_fraction(sample.fraction, replace)
}
\arguments{
\item{sample.fraction}{scalar, fraction of data used for subsampling}

\item{replace}{logical, if bootstrap or subsampling should be used}
}
\value{
sample.fraction
}
\description{
Checks the input data of sample.fraction
}
\keyword{internal}
