% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_type}
\alias{check_type}
\title{check prediction type for predict.orf}
\usage{
check_type(type)
}
\arguments{
\item{type}{string, prediction type for predict.orf}
}
\value{
type
}
\description{
Checks the input data of type in predict.orf
}
\keyword{internal}
