% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_window}
\alias{check_window}
\title{check window size}
\usage{
check_window(window)
}
\arguments{
\item{window}{scalar, share of SD of X used for margins}
}
\value{
window
}
\description{
Checks the input data of window in margins
}
\keyword{internal}
