% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_Y}
\alias{check_Y}
\title{check Y data}
\usage{
check_Y(Y, X)
}
\arguments{
\item{Y}{matrix of input outcomes Y}

\item{X}{matrix of input features X}
}
\value{
Y
}
\description{
Checks the input data as a numeric matrix/vector
}
\keyword{internal}
