% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefstars.R
\name{coefstars}
\alias{coefstars}
\title{Significance Stars for Estimated Coefficients}
\usage{
coefstars(coefs, pvalues)
}
\arguments{
\item{coefs}{matrix of estimated coefficients or fitted values/predictions}

\item{pvalues}{matrix of pvalues for the corresponding coefficients}
}
\value{
a character matrix of coefficients with significance stars
}
\description{
function for creating significance stars for estimated effects which
can be passed into a xtable for printing latex output
}
\examples{
coefs <- matrix(data = c( 0.25, 0.5, 0.75, 1), ncol = 2, nrow = 2)
pvalues <- matrix(data = c( 0.1, 0.04, 0.009, 0.00001), ncol = 2, nrow = 2)
coefstars(coefs, pvalues)

}
