% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_funs.R
\name{get_forest_weights}
\alias{get_forest_weights}
\title{Get Forest Weights}
\usage{
get_forest_weights(forest, honest_data, train_data)
}
\arguments{
\item{forest}{estimated forest object of type ranger}

\item{honest_data}{honest dataframe}

\item{train_data}{train dataframe}
}
\value{
matrix of honest forest weights
}
\description{
get forest weights, i.e. in-sample weights based on honest or train sample
produced by the random forest algorithm as defined in Wager & Athey (2018)
}
\keyword{internal}
