% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/honest_funs.R
\name{get_honest}
\alias{get_honest}
\title{Get Honest Predictions}
\usage{
get_honest(forest, honest_data, train_data)
}
\arguments{
\item{forest}{estimated forest object of type ranger}

\item{honest_data}{honest dataframe}

\item{train_data}{train dataframe}
}
\value{
vector of honest forest predictions
}
\description{
get honest prediction, i.e. fitted values for in sample data
(train and honest sample) based on the honest sample
}
\keyword{internal}
