% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrf_funs.R
\name{get_variance}
\alias{get_variance}
\title{Get Variance}
\usage{
get_variance(honest_pred, honest_weights, honest_outcomes)
}
\arguments{
\item{honest_pred}{vector of honest forest predictions}

\item{honest_weights}{n x n matrix of honest forest weights (based on train)}

\item{honest_outcomes}{vector of real outcomes for the honest sample}
}
\value{
vector of variances
}
\description{
get variance of regression random forest predictions based on honest sample
splitting as described in Lechner (2018)
}
