% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_weights_C}
\alias{get_weights_C}
\title{Get honest weights (C++)}
\usage{
get_weights_C(x, y, z)
}
\arguments{
\item{x}{leaf_IDs_train - list of leaf IDs in train data}

\item{y}{leaf_IDs - list of leaf IDs in honest data}

\item{z}{leaf_size - list of leaf sizes in honest data}
}
\description{
Computes honest weights from the random forest as in Wager & Athey (2019)
for the train and honest sample based on the honest training sample
}
\keyword{internal}
