% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{honest_split}
\alias{honest_split}
\title{honest 50/50 sample split}
\usage{
honest_split(data)
}
\arguments{
\item{data}{dataframe or matrix of features and outcomes to be split honestly}
}
\value{
named list of honest and training sample
}
\description{
Creates honest sample split by randomly selecting 50 percent of observations
to belong to honest sample and 50 percent to training sample
}
