% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margins.R
\name{margins.default}
\alias{margins.default}
\title{margins.default}
\usage{
\method{margins}{default}(forest, eval = NULL, inference = NULL,
  window = NULL, newdata = NULL)
}
\arguments{
\item{forest}{estimated forest object of class \code{orf} or \code{mrf}}

\item{eval}{string defining evaluation point for marginal effects. These can be one of "mean", "atmean", or "atmedian". (Default is "mean")}

\item{inference}{logical, if TRUE inference on marginal effects will be conducted (default is inherited from the orf object)}

\item{window}{numeric, share of standard deviation of X to be used for evaluation of the marginal effect (default is 0.1)}

\item{newdata}{matrix of new Xs for which marginal effects will be computed}
}
\description{
margins.default is a default for S3 generic with methods to estimate marginal effects
of various discrete choice models based on the random forests algorithm.
Applicable classes are \code{orf} or \code{mrf} .#'
}
