% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margins.R
\name{margins.mrf}
\alias{margins.mrf}
\title{margins.mrf}
\usage{
\method{margins}{mrf}(forest, eval = "atmean", newdata = NULL)
}
\arguments{
\item{forest}{trained multinomial random forest object of class \code{mrf}}

\item{eval}{string defining evaluation point for marginal effects. These can be one of "mean", "atmean", or "atmedian". Default is "atmean".}

\item{newdata}{matrix of new Xs (currently not supported)}
}
\value{
object of type \code{margins.mrf}
}
\description{
estimate marginal effects of the multinomial random forest
}
