% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margins.R
\name{margins.orf}
\alias{margins.orf}
\title{margins.orf}
\usage{
\method{margins}{orf}(forest, eval, newdata)
}
\arguments{
\item{forest}{trained ordered random forest object of class \code{orf}}

\item{eval}{string defining evaluation point for marginal effects. These can be one of "mean", "atmean", or "atmedian"}

\item{newdata}{matrix of new Xs (currently not supported)}
}
\value{
object of type \code{margins.orf}
}
\description{
estimate marginal effects of the ordered random forest
}
