% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margins.R
\name{margins}
\alias{margins}
\title{margins}
\usage{
margins(forest, eval, newdata)
}
\arguments{
\item{forest}{trained forest object of class \code{orf}/\code{mrf}/\code{brf}}

\item{eval}{string defining evaluation point for marginal effects. These can be one of "mean", "atmean", or "atmedian"}

\item{newdata}{matrix of new Xs (currently not supported)}
}
\description{
estimate marginal effects of various discrete choice models based on the
random forests algorithm. Applicable classes are \code{orf}, \code{mrf} and
\code{brf}.
}
