% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margins.R
\name{margins_output}
\alias{margins_output}
\title{Formatted output for marginal effects with inference}
\usage{
margins_output(x)
}
\arguments{
\item{x}{object of type \code{margins.orf}}
}
\description{
function for creating inference table output for estimated effects which
can be passed into \code{print.margins.orf}
}
\keyword{internal}
