% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrf_user.R
\name{mrf}
\alias{mrf}
\title{mrf}
\usage{
mrf(X, Y, ntree, mtry, nmin, honesty, inference, margins)
}
\arguments{
\item{X}{matrix of features}

\item{Y}{vector of outcomes (as.matrix acceptable too)}

\item{ntree}{scalar, number of trees in a forest, i.e. bootstrap replications}

\item{mtry}{scalar, number of randomly selected features}

\item{nmin}{scalar, minimum node size}

\item{honesty}{logical, if TRUE honest forest is built using 50:50 data split}

\item{inference}{logical, if TRUE the weight based inference is conducted}

\item{margins}{logical, if TRUE marginal effects at mean are estimated}
}
\value{
object of type mrf
}
\description{
Multinomial Random Forests for semi-parametric estimation of multinomial
choice model as proposed in Lechner & Okasa (2019)
}
