% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrf_dev.R
\name{mrf_margins}
\alias{mrf_margins}
\title{MRF margins}
\usage{
mrf_margins(forest, data, honesty, inference)
}
\arguments{
\item{forest}{list of ranger forest objects}

\item{data}{list of n x n matrices of indicator data for orf}

\item{honesty}{logical if HONESTY should be conducted}

\item{inference}{logical if INFERENCE should be conducted}
}
\value{
matrix of marginal effects estimates
}
\description{
estimate marginal effects for multinomial random forests as defined in
Lechner & Okasa (2019) for in-sample
}
