% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evals.R
\name{mse}
\alias{mse}
\title{Mean Squared Error}
\usage{
mse(predictions, observed)
}
\arguments{
\item{predictions}{matrix of predictions (n x categories)}

\item{observed}{vector of observed ordered categorical outcomes (n x 1)}
}
\value{
scalar, sum MSE for given predictions
}
\description{
computes the mean squared error (MSE) for evaluating the accuracy of
ordered/unordered probability predictions
}
\keyword{internal}
