% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrf_user.R
\name{plot.mrf}
\alias{plot.mrf}
\title{plot.mrf}
\usage{
\method{plot}{mrf}(x, ...)
}
\arguments{
\item{x}{estimated ordered random forest object of type \code{mrf}}

\item{...}{further arguments (currently ignored)}
}
\description{
plot multinomial random forest object of class \code{mrf}
}
