% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrf_user.R
\name{plot.rrf}
\alias{plot.rrf}
\title{plot.rrf}
\usage{
\method{plot}{rrf}(x, ...)
}
\arguments{
\item{x}{estimated forest object of type \code{rrf}}

\item{...}{further arguments (currently ignored)}
}
\description{
plot forest object of class \code{rrf}
}
