% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pred_honest_C}
\alias{pred_honest_C}
\title{Predict honest predictions (C++)}
\usage{
pred_honest_C(x, y, z, w)
}
\arguments{
\item{x}{unique_leaves (List)[ntree]}

\item{y}{honest_y (NumericVector)[nrow]}

\item{z}{honest_leaves (NumericMatrix)[nrow, ntree]}

\item{w}{test_leaves (NumericMatrix)[nrow, ntree]}
}
\description{
Computes honest predictions from the random forest for a test sample based
on the honest training sample
}
\keyword{internal}
