% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrf_dev.R
\name{pred_mrf_margins}
\alias{pred_mrf_margins}
\title{Predict MRF Margins}
\usage{
pred_mrf_margins(forest, data, newdata, honesty, inference)
}
\arguments{
\item{forest}{list of ranger forest objects}

\item{data}{list of n x n matrices of indicator data for orf}

\item{newdata}{matrix of new Xs}

\item{honesty}{logical if HONESTY should be conducted}

\item{inference}{logical if INFERENCE should be conducted}
}
\value{
matrix of marginal effects estimates
}
\description{
estimate marginal effects for multinomial random forests as defined in
Lechner & Okasa (2019) for prediction sample
}
