% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrf_dev.R
\name{pred_mrf_variance}
\alias{pred_mrf_variance}
\title{Predict MRF Variance}
\usage{
pred_mrf_variance(honest_pred, honest_weights, Y_ind_honest)
}
\arguments{
\item{honest_pred}{list of vectors of honest forest predictions}

\item{honest_weights}{list of n x n matrices of honest forest weights}

\item{Y_ind_honest}{list of vectors of 0-1 outcomes for the honest sample}
}
\value{
vector of MRF variances
}
\description{
predict variance of multinomial random forest predictions based on honest sample
splitting as described in Lechner (2018)
}
