% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orf_dev.R
\name{pred_orf_margins}
\alias{pred_orf_margins}
\title{Predict ORF Margins}
\usage{
pred_orf_margins(forest, data, newdata, honesty, inference)
}
\arguments{
\item{forest}{list of ranger forest objects}

\item{data}{list of n x n matrices of indicator data for orf}

\item{newdata}{matrix of new Xs}

\item{honesty}{logical if HONESTY should be conducted}

\item{inference}{logical if INFERENCE should be conducted}
}
\value{
matrix of marginal effects estimates
}
\description{
estimate marginal effects for ordered random forests as defined in
Lechner & Okasa (2019) for prediction sample
}
