% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orf_dev.R
\name{pred_orf_variance}
\alias{pred_orf_variance}
\title{Predict ORF Variance}
\usage{
pred_orf_variance(honest_pred, honest_weights, Y_ind_honest)
}
\arguments{
\item{honest_pred}{list of vectors of honest forest predictions}

\item{honest_weights}{list of n x n matrices of honest forest weights}

\item{Y_ind_honest}{list of vectors of 0-1 outcomes for the honest sample}
}
\value{
vector of ORF variances
}
\description{
predict variance of ordered random forest predictions based on honest sample
splitting as described in Lechner (2018)
}
\keyword{internal}
