% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrf_user.R
\name{predict.mrf}
\alias{predict.mrf}
\title{predict.mrf}
\usage{
\method{predict}{mrf}(object, new_data, ...)
}
\arguments{
\item{object}{estimated forest object of type \code{mrf}}

\item{new_data}{matrix X containing the observations to predict}

\item{...}{further arguments (currently ignored)}
}
\value{
object of class \code{mrf.prediction} with elements
  \tabular{ll}{
      \code{trainForests}  \tab saved forests trained for MRF estimations (inherited from \code{ranger}) \cr
      \code{forestInfo}    \tab info containing forest inputs and data used \cr
      \code{forestPredictions} \tab predicted values \cr
      \code{forestVariances} \tab variances of predicted values (only if \code{inference=TRUE} in the passed \code{mrf object}) \cr
      \code{predictedCategories} \tab categories predicted with MRF based on maximum probability \cr
      \code{marginalEffects} \tab estimated marginal effects from MRF (optionally with inference) \cr
  }
}
\description{
Prediction for new observations based on estimated multinomial random forest
of class \code{mrf}
}
