% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrf_user.R
\name{predict.rrf}
\alias{predict.rrf}
\title{predict.rrf}
\usage{
\method{predict}{rrf}(object, new_data, ...)
}
\arguments{
\item{object}{estimated forest object of type \code{rrf}}

\item{new_data}{matrix X containing the observations to predict}

\item{...}{further arguments (currently ignored)}
}
\value{
object of class \code{rrf.prediction} with elements
  \tabular{ll}{
      \code{forestInfo}    \tab info containing forest inputs and data used \cr
      \code{forestPredictions} \tab predicted values \cr
      \code{forestVariances} \tab variances of predicted values (only if \code{inference=TRUE} in the passed \code{rrf object}) \cr
  }
}
\description{
Prediction for new observations based on estimated forest of type \code{rrf}
}
