% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orf_dev.R
\name{predict_forest_preds_for_ME}
\alias{predict_forest_preds_for_ME}
\title{ORF Predictions for Marginal Effects}
\usage{
predict_forest_preds_for_ME(forest, data, pred_data)
}
\arguments{
\item{forest}{list of ranger forest objects}

\item{data}{list of n x n matrices of indicator data for orf}

\item{pred_data}{list of prediction data (X_mean_up/down)}
}
\value{
list of predictions
}
\description{
Fast ORF Predictions for estimation of marginal effects at mean
}
\keyword{internal}
