% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_funs.R
\name{predict_forest_weights}
\alias{predict_forest_weights}
\title{Predict Forest Weights}
\usage{
predict_forest_weights(forest, data, pred_data)
}
\arguments{
\item{forest}{estimated forest object of type ranger}

\item{data}{train (honest) dataframe}

\item{pred_data}{prediction dataframe}
}
\value{
matrix of honest forest weights
}
\description{
predict forest weights, i.e. out-of-sample weights based on honest or train
sample produced by the random forest algorithm as defined in Wager & Athey (2018)
}
\keyword{internal}
