% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/honest_funs.R
\name{predict_honest}
\alias{predict_honest}
\title{Predict Honest Predictions}
\usage{
predict_honest(forest, honest_data, test_data)
}
\arguments{
\item{forest}{estimated forest object of type ranger}

\item{honest_data}{honest dataframe}

\item{test_data}{test dataframe}
}
\value{
vector of honest forest predictions
}
\description{
predict honest prediction for out of sample data
based on the honest sample
}
\keyword{internal}
