% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margins.R
\name{print.margins.mrf}
\alias{print.margins.mrf}
\title{print.margins.mrf}
\usage{
\method{print}{margins.mrf}(x, latex = FALSE, ...)
}
\arguments{
\item{x}{object of type \code{margins.mrf}}

\item{latex}{logical, if latex output should be generated (\code{default = FALSE})}

\item{...}{further arguments (currently ignored)}
}
\description{
print estimated marginal effects from multinomial random forest of class \code{margins.mrf}
}
