% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margins.R
\name{print.margins.orf}
\alias{print.margins.orf}
\title{print.margins.orf}
\usage{
\method{print}{margins.orf}(x, latex = FALSE, ...)
}
\arguments{
\item{x}{object of type \code{margins.orf}}

\item{latex}{logical, if latex output should be generated (\code{default = FALSE})}

\item{...}{further arguments (currently ignored)}
}
\description{
print estimated marginal effects from ordered random forest of class \code{margins.orf}
}
\examples{
#\dontrun{

## Ordered Forest
require(orf)

# load example data
data(odata)

# specify response and covariates
Y <- odata[, 1]
X <- odata[, -1]

# estimate Ordered Forest
set.seed(123)
orf <- orf(X, Y)

# estimate marginal effects of the orf
orf_margins <- margins(orf)

# print marginal effects
print(orf_margins)

# print marginal effects coded in LaTeX
print(orf_margins, latex = TRUE)

#}

}
