% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orf_user.R
\name{print.orf}
\alias{print.orf}
\title{print.orf}
\usage{
\method{print}{orf}(x, ...)
}
\arguments{
\item{x}{object of type \code{margins.orf}}

\item{...}{further arguments (currently ignored)}
}
\description{
print of an ordered forest object of class \code{orf}
}
