% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margins.R
\name{rep_row}
\alias{rep_row}
\title{repeat rows of a matrix}
\usage{
rep_row(matrix, n)
}
\arguments{
\item{matrix}{matrix which rows should be replicated}

\item{n}{number of times to repeat}
}
\description{
function for replicating rows of a matrix n number of times
}
\keyword{internal}
