% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evals.R
\name{rps}
\alias{rps}
\title{Ranked Probability Score}
\usage{
rps(predictions, observed)
}
\arguments{
\item{predictions}{matrix of predictions (n x categories)}

\item{observed}{vector of observed ordered categorical outcomes (n x 1)}
}
\value{
scalar, mean RPS for given predictions
}
\description{
Computes the mean ranked probability score (RPS) for evaluating the accuracy of
ordered probability predictions
}
\keyword{internal}
