% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrf_user.R
\name{rrf}
\alias{rrf}
\title{rrf}
\usage{
rrf(X, Y, ntree, mtry, nmin, honesty, inference)
}
\arguments{
\item{X}{matrix of features}

\item{Y}{vector of outcomes (as.matrix acceptable too)}

\item{ntree}{scalar, number of trees in a forest, i.e. bootstrap replications}

\item{mtry}{scalar, number of randomly selected features}

\item{nmin}{scalar, minimum node size}

\item{honesty}{logical, if TRUE honest forest is built using 50:50 data split}

\item{inference}{logical, if TRUE the weight based inference is conducted}
}
\value{
object of type rrf
}
\description{
Regression Random Forests (adaptation of ranger with added honesty sample
splitting option, forest weights and inference rpocedure as in Lechner (2018))
}
