% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margins.R
\name{summary.margins.orf}
\alias{summary.margins.orf}
\title{Summary of the Ordered Forest Marginal Effects}
\usage{
\method{summary}{margins.orf}(object, latex = FALSE, ...)
}
\arguments{
\item{object}{object of type \code{margins.orf}}

\item{latex}{logical, if latex output should be generated (\code{default = FALSE})}

\item{...}{further arguments (currently ignored)}
}
\description{
summary of estimated marginal effects of the Ordered Forest of class \code{margins.orf}
}
\details{
\code{summary.margins.orf} provides estimation results of the Ordered Forest
marginal effects. The summary contains the results for the marginal effects
for each covariate and each outcome class, optionally with inference as well.
Furthermore, summary output as a LaTeX table is supported in order to directly
extract the results for the documentation.
}
\examples{
#\dontrun{

## Ordered Forest
require(orf)

# load example data
data(odata)

# specify response and covariates
Y <- odata[, 1]
X <- odata[, -1]

# estimate Ordered Forest
set.seed(123)
orf <- orf(X, Y)

# estimate marginal effects of the orf
orf_margins <- margins(orf)

# summary of marginal effects
summary(orf_margins)

# summary of marginal effects coded in LaTeX
summary(orf_margins, latex = TRUE)

#}

}
