% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrf_user.R
\name{summary.mrf}
\alias{summary.mrf}
\title{summary.mrf}
\usage{
\method{summary}{mrf}(object, latex, ...)
}
\arguments{
\item{object}{estimated multinomial random forest object of class \code{mrf}}

\item{latex}{logical, TRUE if latex summary should be generated}

\item{...}{further arguments (currently ignored)}
}
\description{
summary of an multinomial random forest object of class \code{mrf}
}
