% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orf_user.R
\name{summary.orf.prediction}
\alias{summary.orf.prediction}
\title{orf prediction summary}
\usage{
\method{summary}{orf.prediction}(object, latex = FALSE, ...)
}
\arguments{
\item{object}{predicted Ordered Forest object of class \code{orf.prediction}}

\item{latex}{logical, TRUE if latex summary should be generated}

\item{...}{further arguments (currently ignored)}
}
\description{
summary of Ordered Forest predictions of class \code{orf.prediction}
}
\details{
\code{summary.orf.prediction} provides a main summary of the Ordered Forest
prediction, including the input information regarding the values of hyperparameters
as well as the inputs of the \code{predict.orf} function.
}
\examples{
#\dontrun{

# Ordered Forest
require(orf)

# load example data
data(odata)

# specify response and covariates for train and test
set.seed(123)
idx <- sample(seq(1, nrow(odata), 1), 0.8*nrow(odata))

# train set
Y_train <- odata[idx, 1]
X_train <- odata[idx, -1]

# test set
Y_test <- odata[-idx, 1]
X_test <- odata[-idx, -1]

# estimate Ordered Forest
set.seed(123)
orf <- orf(X_train, Y_train)

# predict the probabilities with the estimated orf
orf_pred <- predict(orf, newdata = X_test)

# summary of the prediction object
summary(orf_pred)

#}


}
