% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrf_user.R
\name{summary.rrf}
\alias{summary.rrf}
\title{summary.rrf}
\usage{
\method{summary}{rrf}(object, latex, ...)
}
\arguments{
\item{object}{estimated forest object of type \code{rrf}}

\item{latex}{logical, TRUE if latex summary should be generated}

\item{...}{further arguments (currently ignored)}
}
\description{
summary of a forest object of class \code{rrf}
}
