% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margins.R
\name{margins.default}
\alias{margins.default}
\title{margins.default}
\usage{
\method{margins}{default}(forest, eval, newdata)
}
\arguments{
\item{forest}{estimated forest object of class \code{orf} or \code{mrf}}

\item{eval}{string defining evaluation point for marginal effects. These can be one of "mean", "atmean", or "atmedian"}

\item{newdata}{matrix of new Xs (currently not supported)}
}
\description{
margins.default is a default for S3 generic with methods to estimate marginal effects
of various discrete choice models based on the random forests algorithm.
Applicable classes are \code{orf} or \code{mrf} .#'
}
