% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margins.R
\name{margins.mrf}
\alias{margins.mrf}
\title{margins.mrf}
\usage{
\method{margins}{mrf}(forest, eval = NULL, inference = NULL,
  window = NULL, newdata = NULL)
}
\arguments{
\item{forest}{trained multinomial random forest object of class \code{mrf}}

\item{eval}{string defining evaluation point for marginal effects. These can be one of "mean", "atmean", or "atmedian". (Default is "mean")}

\item{inference}{logical, if TRUE inference on marginal effects will be conducted (default is inherited from the mrf object)}

\item{window}{numeric, share of standard deviation of X to be used for evaluation of the marginal effect (default is 0.1)}

\item{newdata}{matrix of new Xs for which marginal effects will be computed}
}
\value{
object of type \code{margins.mrf}
}
\description{
estimate marginal effects of the multinomial random forest
}
