% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margins.R
\name{margins.orf}
\alias{margins.orf}
\title{margins.orf}
\usage{
\method{margins}{orf}(forest, eval = NULL, inference = NULL,
  window = NULL, newdata = NULL)
}
\arguments{
\item{forest}{trained ordered random forest object of class \code{orf}}

\item{eval}{string defining evaluation point for marginal effects. These can be one of "mean", "atmean", or "atmedian" (default is "mean")}

\item{inference}{logical, if TRUE inference on marginal effects will be conducted (default is inherited from the orf object)}

\item{window}{numeric, share of standard deviation of X to be used for evaluation of the marginal effect (default is 0.1)}

\item{newdata}{matrix of new Xs for which marginal effects will be computed}
}
\value{
object of type \code{margins.orf}
}
\description{
estimate marginal effects of the ordered random forest
}
\examples{
#\dontrun{

## Ordered Forest
require(orf)

# load example data
data(odata)

# specify response and covariates
Y <- odata[, 1]
X <- odata[, -1]

# estimate Ordered Forest
set.seed(123)
orf <- orf(X, Y)

# estimate marginal effects of the orf (default)
margins(orf)

# estimate marginal effects evaluated at the mean
margins(orf, eval = "atmean")

# estimate marginal effects with inference
# (orf object has to be estimated with honesty and subsampling)
margins(orf, inference = TRUE)

# estimate marginal effects with custom window size
margins(orf, window = 0.5)

# estimate marginal effects for some new data (within support of X)
margins(orf, newdata = X[1:10, ])

# estimate marginal effects with all custom settings
margins(orf, eval = "atmedian", inference = TRUE, window = 0.5, newdata = X[1:10, ])

#}

}
