% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margins.R
\name{margins}
\alias{margins}
\title{margins}
\usage{
margins(forest, eval = "atmean", window = NULL, newdata = NULL)
}
\arguments{
\item{forest}{estimated forest object of class \code{orf} or \code{mrf}}

\item{eval}{string defining evaluation point for marginal effects. These can be one of "mean", "atmean", or "atmedian"}

\item{window}{numeric, share of standard deviation of X to be used for evaluation of the marginal effect (default is 0.1)}

\item{newdata}{matrix of new Xs for which marginal effects will be computed}
}
\description{
margins is an S3 generic with methods to estimate marginal effects
of various discrete choice models based on the random forests algorithm.
Applicable classes are \code{orf} and \code{mrf}.
}
