% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orf_user.R
\name{plot.orf}
\alias{plot.orf}
\title{plot.orf}
\usage{
\method{plot}{orf}(x, ...)
}
\arguments{
\item{x}{estimated ordered random forest object of type \code{orf}}

\item{...}{further arguments (currently ignored)}
}
\description{
plot ordered random forest object of class \code{orf}
}
\examples{
#\dontrun{

# Ordered Forest
require(orf)

# load example data
data(odata)

# specify response and covariates
Y <- odata[, 1]
X <- odata[, -1]

# estimate Ordered Forest
set.seed(123)
orf <- orf(X, Y)

# plot the estimated probability distributions
plot(orf)

#}

}
