% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orf_user.R
\name{predict.orf}
\alias{predict.orf}
\title{predict.orf}
\usage{
\method{predict}{orf}(object, newdata = NULL, inference = NULL, ...)
}
\arguments{
\item{object}{estimated forest object of type \code{orf}}

\item{newdata}{matrix X containing the observations to predict}

\item{inference}{logical, if TRUE variances for the predictions will be estimated}

\item{...}{further arguments (currently ignored)}
}
\value{
object of class \code{orf.prediction} with elements
  \tabular{ll}{
      \code{trainForests}  \tab saved forests trained for ORF estimations (inherited from \code{ranger}) \cr
      \code{forestInfo}    \tab info containing forest inputs and data used \cr
      \code{forestPredictions} \tab predicted values \cr
      \code{forestVariances} \tab variances of predicted values (only if \code{inference=TRUE} in the passed \code{orf object}) \cr
      \code{predictedCategories} \tab categories predicted with ORF based on maximum probability \cr
  }
}
\description{
Prediction for new observations based on estimated ordered random forest
of type \code{orf}
}
