% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orf_user.R
\name{predict.orf}
\alias{predict.orf}
\title{predict.orf}
\usage{
\method{predict}{orf}(object, new_data, ...)
}
\arguments{
\item{object}{estimated forest object of type \code{rrf}}

\item{new_data}{matrix X containing the observations to predict}

\item{...}{further arguments (currently ignored)}
}
\value{
object of class \code{orf.prediction} with elements
  \tabular{ll}{
  "trainForests",  "forestInfo", "honestPredictions", "honestVariances", "predictedCategories", "marginalEffects")
      \code{trainForests}  \tab saved forests trained for ORF estimations (inherited from \code{ranger}) \cr
      \code{forestInfo}    \tab info containing forest inputs and data used \cr
      \code{forestPredictions} \tab predicted values \cr
      \code{forestVariances} \tab variances of predicted values (only if \code{inference=TRUE} in the passed \code{orf object}) \cr
      \code{predictedCategories} \tab categories predicted with ORF based on maximum probability \cr
      \code{marginalEffects} \tab estimated marginal effects from ORF (optionally with inference) \cr
  }
}
\description{
Prediction for new observations based on estimated ordered random forest
of type \code{orf}
}
