% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margins.R
\name{print.margins.orf}
\alias{print.margins.orf}
\title{print.margins.orf}
\usage{
\method{print}{margins.orf}(x, latex = FALSE, ...)
}
\arguments{
\item{x}{object of type \code{margins.orf}}

\item{latex}{logical, if latex output should be generated (\code{default = FALSE})}

\item{...}{further arguments (currently ignored)}
}
\description{
print estimated marginal effects from ordered random forest of class \code{orf}
}
