% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orf_user.R
\name{print.orf}
\alias{print.orf}
\title{print.orf}
\usage{
\method{print}{orf}(x, ...)
}
\arguments{
\item{x}{object of type \code{orf}}

\item{...}{further arguments (currently ignored)}
}
\description{
print of an ordered forest object of class \code{orf}
}
\examples{
#\dontrun{

# Ordered Forest
require(orf)

# load example data
data(odata)

# specify response and covariates
Y <- odata[, 1]
X <- odata[, -1]

# estimate Ordered Forest
set.seed(123)
orf <- orf(X, Y)

# print output of the orf estimation
print(orf)

#}


}
