% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orf_user.R
\name{summary.orf}
\alias{summary.orf}
\title{summary.orf}
\usage{
\method{summary}{orf}(object, latex, ...)
}
\arguments{
\item{object}{estimated ordered random forest object of type \code{orf}}

\item{latex}{logical, TRUE if latex summary should be generated}

\item{...}{further arguments (currently ignored)}
}
\description{
summary of an ordered random forest object of class \code{orf}
}
